Move casting_utils to LibASR
So it can be used in intrinsic_function_registry
